+++ swarm/src/behaviour/toggle.rs
@@ -418,7 +418,7 @@ where
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::handler::DummyConnectionHandler;
+    use crate::dummy;
 
     /// A disabled [`ToggleConnectionHandler`] can receive listen upgrade errors in
     /// the following two cases:
@@ -436,7 +436,7 @@ mod tests {
     /// [`ToggleConnectionHandler`] should ignore the error in both of these cases.
     #[test]
     fn ignore_listen_upgrade_error_when_disabled() {
-        let mut handler = ToggleConnectionHandler::<DummyConnectionHandler> { inner: None };
+        let mut handler = ToggleConnectionHandler::<dummy::ConnectionHandler> { inner: None };
 
         handler.inject_listen_upgrade_error(Either::Right(()), ConnectionHandlerUpgrErr::Timeout);
     }